// src/app/shared/roles.enum.ts
export enum Role {
    joborder = 'JOBORDER',
    jobordereditor = 'JOBORDEREDITOR',
    joborderadmin = 'JOBORDERADMIN',
    invoice = 'INVOICE',
    invoiceeditor = 'INVOICEEDITOR',
    invoiceadmin = 'INVOICEADMIN',
    receipt = 'RECEIPT',
    receipteditor = 'RECEIPTEDITOR',
    receiptadmin = 'RECEIPTADMIN',
    report = 'REPORT',
    inquiry = 'INQUIRY',
    admin = 'ADMIN'
}